deduplicate effect choices
