+++ meerk40t/moshi/device.py
@@ -14,6 +14,7 @@ from ..core.units import Length
 from ..device.mixins import Status
 from .controller import MoshiController
 from .driver import MoshiDriver
+from meerk40t.device.devicechoices import get_effect_choices
 
 
 class MoshiDevice(Service, Status):
@@ -222,54 +223,7 @@ class MoshiDevice(Service, Status):
         ]
         self.register_choices("coolant", choices)
 
-        choices = [
-            {
-                "attr": "effect_hatch_default_distance",
-                "object": self,
-                "default": "1.0mm",
-                "type": str,
-                "label": _("Hatch Distance"),
-                "tip": _("Default Hatch Distance"),
-                "section": "Effect Defaults",
-            },
-            {
-                "attr": "effect_hatch_default_angle",
-                "object": self,
-                "default": "0deg",
-                "type": str,
-                "label": _("Hatch Angle"),
-                "tip": _("Default Hatch Angle"),
-                "section": "Effect Defaults",
-            },
-            {
-                "attr": "effect_hatch_default_angle_delta",
-                "object": self,
-                "default": "0deg",
-                "type": str,
-                "label": _("Hatch Angle Delta"),
-                "tip": _("Default Hatch Angle Delta"),
-                "section": "Effect Defaults",
-            },
-            {
-                "attr": "effect_wobble_default_radius",
-                "object": self,
-                "default": "0.5mm",
-                "type": str,
-                "label": _("Wobble Radius"),
-                "tip": _("Default Wobble Radius"),
-                "section": "Effect Defaults",
-            },
-            {
-                "attr": "effect_wobble_default_interval",
-                "object": self,
-                "default": "0.05mm",
-                "type": str,
-                "label": _("Wobble Interval"),
-                "tip": _("Default Wobble Interval"),
-                "section": "Effect Defaults",
-            },
-        ]
-        self.register_choices("moshi-effects", choices)
+        self.register_choices("moshi-effects", get_effect_choices(self))
 
         # Tuple contains 4 value pairs: Speed Low, Speed High, Power Low, Power High, each with enabled, value
         self.setting(
