+++ meerk40t/newly/device.py
@@ -8,6 +8,7 @@ from meerk40t.core.view import View
 from meerk40t.device.mixins import Status
 from meerk40t.kernel import CommandSyntaxError, Service, signal_listener
 from meerk40t.newly.driver import NewlyDriver
+from meerk40t.device.devicechoices import get_effect_choices
 
 
 class NewlyDevice(Service, Status):
@@ -106,54 +107,7 @@ class NewlyDevice(Service, Status):
         ]
         self.register_choices("newly-speedchart", choices)
 
-        choices = [
-            {
-                "attr": "effect_hatch_default_distance",
-                "object": self,
-                "default": "1.0mm",
-                "type": str,
-                "label": _("Hatch Distance"),
-                "tip": _("Default Hatch Distance"),
-                "section": "Effect Defaults",
-            },
-            {
-                "attr": "effect_hatch_default_angle",
-                "object": self,
-                "default": "0deg",
-                "type": str,
-                "label": _("Hatch Angle"),
-                "tip": _("Default Hatch Angle"),
-                "section": "Effect Defaults",
-            },
-            {
-                "attr": "effect_hatch_default_angle_delta",
-                "object": self,
-                "default": "0deg",
-                "type": str,
-                "label": _("Hatch Angle Delta"),
-                "tip": _("Default Hatch Angle Delta"),
-                "section": "Effect Defaults",
-            },
-            {
-                "attr": "effect_wobble_default_radius",
-                "object": self,
-                "default": "0.5mm",
-                "type": str,
-                "label": _("Wobble Radius"),
-                "tip": _("Default Wobble Radius"),
-                "section": "Effect Defaults",
-            },
-            {
-                "attr": "effect_wobble_default_interval",
-                "object": self,
-                "default": "0.05mm",
-                "type": str,
-                "label": _("Wobble Interval"),
-                "tip": _("Default Wobble Interval"),
-                "section": "Effect Defaults",
-            },
-        ]
-        self.register_choices("newly-effects", choices)
+        self.register_choices("newly-effects", get_effect_choices(self))
 
         choices = [
             {
